Fix integrationTestTask execution problem ...
So far, just running the `test` task will run Tomcat automatically,

There is a problem that the server is executed just by running the mock test,
which does not require server execution, so I made the following changes.

1. Run `gradle test` to run only the test without setting the `integration` tag.
2. Run `gradle integrationTest` to run only tests with the `integration` tag set.

---

Fix of the method for Gretty to obtain a random port

The method of obtaining a random port from the servicePort value in the gretty_ports.properties file was incorrect.

The following code can be used to obtain a random port.:
System.getProperty("gretty.httpPort")

* https://akhikhl.github.io/gretty-doc/Integration-tests-support.html